Moved createAbstract out of Prepack global instrinct to make it more re-usable
